+++ pears/templates/results.html
@@ -73,7 +73,8 @@
       <div class="row">
         <div class="col-lg-12">
           <h3><a href="{{result[0]}}">{{result[1]}}</a></h3>
-          {{result[2]}} ... <a href="{{result[3]}}">Cache</a>
+          {{result[2]}} ... <a href="{{result[3]}}">Cache</a><br>
+          <a href="{{result[0]}}">{{result[0]}}</a>
           <hr/>
         </div>
       </div>
